requirements.txt
# KAVACH – Backend Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pyzbar>=0.1.9
python-whois>=0.9.4
requests>=2.31.0
pytesseract>=0.3.10
scikit-learn>=1.3.0
easyocr>=1.7.1
pillow>=10.0.0
opencv-python>=4.8.0
tensorflow>=2.15.0
